Move framework header into cpp
