+++ bcos-executor/src/precompiled/common/WorkingSealerManagerImpl.cpp
@@ -28,6 +28,7 @@
 #include "bcos-framework/sealer/VrfCurveType.h"
 #include "fmt/ranges.h"
 #include "bcos-framework/ledger/LedgerTypeDef.h"
+#include <fmt/compile.h>
 #include <fmt/format.h>
 #include <boost/endian/conversion.hpp>
 #include <algorithm>
