+++ bcos-framework/bcos-framework/protocol/Transaction.h
@@ -18,19 +18,15 @@
  */
 #pragma once
 #include "TransactionSubmitResult.h"
-#include <bcos-crypto/interfaces/crypto/CryptoSuite.h>
 #include <bcos-crypto/interfaces/crypto/Hash.h>
-#include <bcos-crypto/interfaces/crypto/KeyInterface.h>
-#include <bcos-utilities/BoostLog.h>
+#include <bcos-crypto/interfaces/crypto/Signature.h>
 #include <bcos-utilities/Common.h>
 #include <bcos-utilities/Error.h>
-#include <range/v3/view/any_view.hpp>
-#if !ONLY_CPP_SDK
-#include <bcos-utilities/ITTAPI.h>
-#endif
 #include "bcos-utilities/AnyHolder.h"
-#include <bcos-crypto/hash/Keccak256.h>
-#include <boost/throw_exception.hpp>
+#include <atomic>
+#include <compare>
+#include <concepts>
+#include <functional>
 #include <utility>
 
 namespace bcostars::protocol
@@ -79,62 +75,10 @@ public:
     using ConstPtr = std::shared_ptr<const Transaction>;
 
     Transaction() = default;
-    Transaction(const Transaction& other)
-      : m_submitCallback(other.m_submitCallback),
-        m_batchHash(other.m_batchHash),
-        m_batchId(other.m_batchId),
-        m_synced(other.m_synced),
-        m_sealed(other.m_sealed),
-        m_invalid(other.m_invalid),
-        m_systemTx(other.m_systemTx.load(std::memory_order_acquire)),
-                m_tainted(other.m_tainted),
-        m_storeToBackend(other.m_storeToBackend)
-    {}
-    Transaction(Transaction&& other) noexcept
-      : m_submitCallback(std::move(other.m_submitCallback)),
-        m_batchHash(other.m_batchHash),
-        m_batchId(other.m_batchId),
-        m_synced(other.m_synced),
-        m_sealed(other.m_sealed),
-        m_invalid(other.m_invalid),
-        m_systemTx(other.m_systemTx.load(std::memory_order_acquire)),
-                m_tainted(other.m_tainted),
-        m_storeToBackend(other.m_storeToBackend)
-    {}
-    Transaction& operator=(const Transaction& other)
-    {
-        if (this != &other)
-        {
-            m_submitCallback = other.m_submitCallback;
-            m_batchHash = other.m_batchHash;
-            m_batchId = other.m_batchId;
-            m_synced = other.m_synced;
-            m_sealed = other.m_sealed;
-            m_invalid = other.m_invalid;
-            m_systemTx.store(
-                other.m_systemTx.load(std::memory_order_acquire), std::memory_order_release);
-            m_tainted = other.m_tainted;
-            m_storeToBackend = other.m_storeToBackend;
-        }
-        return *this;
-    }
-    Transaction& operator=(Transaction&& other) noexcept
-    {
-        if (this != &other)
-        {
-            m_submitCallback = std::move(other.m_submitCallback);
-            m_batchHash = other.m_batchHash;
-            m_batchId = other.m_batchId;
-            m_synced = other.m_synced;
-            m_sealed = other.m_sealed;
-            m_invalid = other.m_invalid;
-            m_systemTx.store(
-                other.m_systemTx.load(std::memory_order_acquire), std::memory_order_release);
-            m_tainted = other.m_tainted;
-            m_storeToBackend = other.m_storeToBackend;
-        }
-        return *this;
-    }
+    Transaction(const Transaction& other);
+    Transaction(Transaction&& other) noexcept;
+    Transaction& operator=(const Transaction& other);
+    Transaction& operator=(Transaction&& other) noexcept;
     virtual ~Transaction() noexcept = default;
 
     virtual void decode(bytesConstRef _txData) = 0;
@@ -142,44 +86,7 @@ public:
     virtual bcos::crypto::HashType hash() const = 0;
     virtual bcos::bytesConstRef extraTransactionBytes() const = 0;
 
-    virtual void verify(crypto::Hash& hashImpl, crypto::SignatureCrypto& signatureImpl)
-    {
-#if !ONLY_CPP_SDK
-        ittapi::Report report(ittapi::ITT_DOMAINS::instance().TRANSACTION,
-            ittapi::ITT_DOMAINS::instance().VERIFY_TRANSACTION);
-#endif
-        // The tx has already been verified
-        if (!tainted())
-        {
-            return;
-        }
-        // based on type to switch recover sender
-        crypto::HashType hashResult;
-        if (type() == static_cast<uint8_t>(TransactionType::BCOSTransaction))
-        {
-            // Always recompute hash from transaction data to prevent content-hash
-            // mismatch attacks where the embedded dataHash field is tampered
-            calculateHash(hashImpl);
-            hashResult = hash();
-        }
-        else if (type() == static_cast<uint8_t>(TransactionType::Web3Transaction))
-        {
-            const auto bytesRef = extraTransactionBytes();
-            hashResult = bcos::crypto::keccak256Hash(bytesRef);
-        }
-        // check the signatures
-        const auto signature = signatureData();
-        auto [recovered, sender] = signatureImpl.recoverAddress(hashImpl, hashResult, signature);
-        if (!recovered) [[unlikely]]
-        {
-            BCOS_LOG(INFO) << LOG_DESC("recover sender address failed")
-                           << LOG_KV("hash", hashResult.abridged());
-            BOOST_THROW_EXCEPTION(
-                std::invalid_argument("recover sender address from signature failed"));
-        }
-        forceSender(sender);
-        setTainted(false);
-    }
+    virtual void verify(crypto::Hash& hashImpl, crypto::SignatureCrypto& signatureImpl);
 
     virtual int32_t version() const = 0;
     virtual std::string_view chainId() const = 0;
@@ -291,30 +198,6 @@ using AnyTransaction =
     AnyHolder<bcos::protocol::Transaction, 184>;  // 多平台TransactinImpl的最大尺寸 (Maximum size of
                                                   // TransactinImpl across platforms)
 
-inline std::ostream& operator<<(std::ostream& stream, const Transaction& transaction)
-{
-    stream << "Transaction{" << "hash=" << transaction.hash() << ", "
-           << "version=" << transaction.version() << ", " << "chainId=" << transaction.chainId()
-           << ", " << "groupId=" << transaction.groupId() << ", "
-           << "blockLimit=" << transaction.blockLimit() << ", " << "nonce=" << transaction.nonce()
-           << ", " << "to=" << transaction.to() << ", " << "abi=" << transaction.abi() << ", "
-           << "value=" << transaction.value() << ", " << "gasPrice=" << transaction.gasPrice()
-           << ", " << "gasLimit=" << transaction.gasLimit() << ", "
-           << "maxFeePerGas=" << transaction.maxFeePerGas() << ", "
-           << "maxPriorityFeePerGas=" << transaction.maxPriorityFeePerGas() << ", "
-           << "extension=" << toHex(transaction.extension()) << ", "
-           << "extraData=" << transaction.extraData() << ", "
-           << "sender=" <<
-        [&]() {
-            auto view = transaction.sender();
-            return bcos::bytesConstRef{(const bcos::byte*)view.data(), view.size()};
-        }() << ", "
-           << "input=" << toHex(transaction.input()) << ", "
-           << "importTime=" << transaction.importTime() << ", "
-           << "type=" << static_cast<int>(transaction.type()) << ", "
-           << "attribute=" << transaction.attribute() << ", "
-           << "size=" << transaction.size() << "}";
-    return stream;
-}
+std::ostream& operator<<(std::ostream& stream, const Transaction& transaction);
 
 }  // namespace bcos::protocol
