+++ bcos-framework/bcos-framework/storage/Entry.h
@@ -2,13 +2,12 @@
 
 #include "Common.h"
 #include "bcos-crypto/interfaces/crypto/Hash.h"
-#include "bcos-framework/protocol/Protocol.h"
 #include <bcos-utilities/Common.h>
 #include <bcos-utilities/Error.h>
 #include <boost/archive/basic_archive.hpp>
-#include <boost/exception/diagnostic_information.hpp>
 #include <boost/iostreams/device/back_inserter.hpp>
 #include <boost/iostreams/stream.hpp>
+#include <boost/lexical_cast.hpp>
 #include <boost/throw_exception.hpp>
 #include <algorithm>
 #include <cstdint>
@@ -96,17 +95,7 @@ public:
 
     std::string_view get() const& { return outputValueView(m_value); }
 
-    std::string_view getField(size_t index) const&
-    {
-        if (index > 0)
-        {
-            BOOST_THROW_EXCEPTION(
-                BCOS_ERROR(-1, "Get field index: " + boost::lexical_cast<std::string>(index) +
-                                   " failed, index out of range"));
-        }
-
-        return get();
-    }
+    std::string_view getField(size_t index) const&;
 
     template <typename T>
     void setField(size_t index, T&& input)
@@ -121,11 +110,7 @@ public:
         set(std::forward<T>(input));
     }
 
-    void set(const char* pointer)
-    {
-        auto view = std::string_view(pointer, strlen(pointer));
-        set(view);
-    }
+    void set(const char* pointer);
 
     void set(EntryBufferInput auto value)
     {
@@ -179,15 +164,7 @@ public:
 
     Status status() const { return m_status; }
 
-    void setStatus(Status status)
-    {
-        m_status = status;
-        if (m_status == DELETED)
-        {
-            m_size = 0;
-            m_value = std::string();
-        }
-    }
+    void setStatus(Status status);
 
     bool dirty() const { return (m_status == MODIFIED || m_status == DELETED); }
 
@@ -209,96 +186,15 @@ public:
         return std::move(m_value);
     }
 
-    const char* data() const&
-    {
-        auto view = outputValueView(m_value);
-        return view.data();
-    }
+    const char* data() const&;
     int32_t size() const { return m_size; }
 
     bool valid() const { return m_status == Status::NORMAL; }
     crypto::HashType hash(std::string_view table, std::string_view key,
-        const bcos::crypto::Hash& hashImpl, uint32_t blockVersion) const
-    {
-        bcos::crypto::HashType entryHash(0);
-        if (blockVersion >= (uint32_t)bcos::protocol::BlockVersion::V3_1_VERSION)
-        {
-            auto hasher = hashImpl.hasher();
-            hasher.update(table);
-            hasher.update(key);
-
-            switch (m_status)
-            {
-            case MODIFIED:
-            {
-                auto data = get();
-                hasher.update(data);
-                hasher.final(entryHash);
-                if (c_fileLogLevel == TRACE) [[unlikely]]
-                {
-                    STORAGE_LOG(TRACE)
-                        << "Entry hash, dirty entry: " << table << " | " << toHex(key) << " | "
-                        << toHex(data) << LOG_KV("hash", entryHash.abridged());
-                }
-                break;
-            }
-            case DELETED:
-            {
-                hasher.final(entryHash);
-                if (c_fileLogLevel == TRACE) [[unlikely]]
-                {
-                    STORAGE_LOG(TRACE) << "Entry hash, deleted entry: " << table << " | "
-                                       << toHex(key) << LOG_KV("hash", entryHash.abridged());
-                }
-                break;
-            }
-            default:
-            {
-                STORAGE_LOG(DEBUG) << "Entry hash, clean entry: " << table << " | " << toHex(key)
-                                   << " | " << (int)m_status;
-                break;
-            }
-            }
-        }
-        else
-        {  // 3.0.0
-            if (m_status == Entry::MODIFIED)
-            {
-                auto value = get();
-                bcos::bytesConstRef ref((const bcos::byte*)value.data(), value.size());
-                entryHash = hashImpl.hash(ref);
-                if (c_fileLogLevel == TRACE) [[unlikely]]
-                {
-                    STORAGE_LOG(TRACE)
-                        << "Entry Calc hash, dirty entry: " << table << " | " << toHex(key) << " | "
-                        << toHex(value) << LOG_KV("hash", entryHash.abridged());
-                }
-            }
-            else if (m_status == Entry::DELETED)
-            {
-                entryHash = bcos::crypto::HashType(0x1);
-                if (c_fileLogLevel == TRACE) [[unlikely]]
-                {
-                    STORAGE_LOG(TRACE) << "Entry Calc hash, deleted entry: " << table << " | "
-                                       << toHex(key) << LOG_KV("hash", entryHash.abridged());
-                }
-            }
-        }
-        return entryHash;
-    }
+        const bcos::crypto::Hash& hashImpl, uint32_t blockVersion) const;
 
 private:
-    [[nodiscard]] auto outputValueView(const ValueType& value) const& -> std::string_view
-    {
-        std::string_view view;
-        std::visit(
-            [this, &view](auto&& valueInside) {
-                auto viewRaw = inputValueView(valueInside);
-                view = std::string_view(viewRaw.data(), m_size);
-            },
-            value);
-        return view;
-    }
+    [[nodiscard]] auto outputValueView(const ValueType& value) const& -> std::string_view;
 
     template <typename T>
     [[nodiscard]] auto inputValueView(const T& value) const -> std::string_view
